// src/generated/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: any; output: any; }
};
// ...
export type CommentCount = {
  __typename?: 'CommentCount';
  hideIfNoComments?: Maybe<Scalars['Boolean']['output']>;
  postId?: Maybe<Scalars['String']['output']>;
  threadId?: Maybe<Scalars['String']['output']>;
};
// ...
export type News = {
  __typename?: 'News';
  cards?: Maybe<Array<Maybe<NewsItem>>>;
};

export type NewsImage = {
  __typename?: 'NewsImage';
  alt?: Maybe<Scalars['String']['output']>;
  height?: Maybe<Scalars['Int']['output']>;
  src?: Maybe<Scalars['String']['output']>;
  srcSet?: Maybe<Array<Maybe<Scalars['Int']['output']>>>;
  width?: Maybe<Scalars['Int']['output']>;
};

export type NewsImageContainer = {
  __typename?: 'NewsImageContainer';
  image?: Maybe<NewsImage>;
};

export type NewsItem = {
  __typename?: 'NewsItem';
  commentCount?: Maybe<CommentCount>;
  image?: Maybe<NewsImageContainer>;
  newsTab?: Maybe<Scalars['Boolean']['output']>;
  publishTime?: Maybe<Scalars['String']['output']>;
  tags?: Maybe<Tags>;
  teaser?: Maybe<Scalars['String']['output']>;
  title?: Maybe<Scalars['String']['output']>;
  url?: Maybe<Scalars['String']['output']>;
};
// ...
export type Tags = {
  __typename?: 'Tags';
  tags?: Maybe<Array<Maybe<Tag>>>;
};